Trying to push changes after i have deleted orginal project from my hdd coz of the bug i couldn't resolve until now.
I thing i am going to make it this time.

I finish my first full stack project, doing fe and be at the same time.
It was hard, especially the fe, but i manage to work it to the end.
I have learned a lot. There are few more thing i have to do to complete
this project like add page for terms and conditions, about us, and stuff
like that.